requirements.txt
certifi==2024.12.14
setuptools==65.5.0
cryptography==44.0.0
pip==24.3.1
azure-storage-blob==12.6.0
click==8.1.8
pyodbc==5.2.0
typing-extensions==4.12.2
idna==3.10
charset-normalizer==3.4.1
oauthlib==3.2.2
wtforms==3.2.1
jinja2==3.1.5
werkzeug==3.1.3
flask-sqlalchemy==3.1.1
greenlet==3.1.1
flask==3.1.0
markupsafe==3.0.2
requests==2.32.3
pycparser==2.22
pyjwt==2.10.1
urllib3==2.3.0
itsdangerous==2.2.0
sqlalchemy==2.0.36
requests-oauthlib==2.0.0
azure-core==1.32.0
msal==1.31.1
cffi==1.17.1
six==1.17.0
blinker==1.9.0
flask-wtf==1.2.2
python-dotenv==1.0.1
msgspec==0.19.0
cachelib==0.13.0
flask-session==0.8.0
isodate==0.7.2
msrest==0.7.1
flask-login==0.6.3
config==0.5.1
colorama==0.4.6
